Added test cases for delete function
testcases for single and multiple deletes successful

Of note, my update function uses the upsert command by supabase. This will update existing entry, or create new one if it doesnt exist. The reason why i have two different functions for create and update is because upsert does not have auto generation of id.

Because of the nature of upsert, it runs asych with the delete function, and the item entries may remain even after the delete test cases are successful. Hence, the teardownclass function is there to run after all testcases have been completed, to clean up the database